refactor(core): reorg organization queries
